Added relationships

+++ src/idea/idea.service.ts
@@ -2,7 +2,7 @@ import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { UserEntity } from '../user/user.entity';
-import { IdeaDTO } from './idea.dto';
+import { IdeaDTO, IdeaRO } from './idea.dto';
 import { IdeaEntity } from './idea.entity';
 
 @Injectable()
@@ -14,43 +14,73 @@ export class IdeaService {
     private readonly userRepository: Repository<UserEntity>,
   ) {}
 
-  async showAll() {
-    return await this.ideaRepository.find({ relations: ['author'] });
-  }
-
-  async create(userId: string, data: IdeaDTO) {
-    const user = await this.userRepository.findOne({ where: { id: userId } });
-    const idea = await this.ideaRepository.create({ ...data, author: user });
-    await this.ideaRepository.save(idea);
+  private toResponseObject(idea: IdeaEntity): IdeaRO {
     return {
       ...idea,
-      author: idea.author.toResponseObject(),
+      author: idea.author.toResponseObject(false),
     };
   }
 
-  async read(id: string) {
-    const idea = await this.ideaRepository.findOne({ where: { id } });
-    if (!idea) {
-      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+  private ensureOwnership(idea: IdeaEntity, userId: string) {
+    if (idea.author.id !== userId) {
+      throw new HttpException('Incorrect User', HttpStatus.UNAUTHORIZED);
     }
-    return idea;
   }
 
-  async update(id: string, data: Partial<IdeaDTO>) {
-    const idea = await this.ideaRepository.findOne({ where: { id } });
+  async showAll(): Promise<IdeaRO[]> {
+    const ideas = await this.ideaRepository.find({ relations: ['author'] });
+    return ideas.map(idea => this.toResponseObject(idea));
+  }
+
+  async create(userId: string, data: IdeaDTO): Promise<IdeaRO> {
+    const user = await this.userRepository.findOne({ where: { id: userId } });
+    const idea = await this.ideaRepository.create({ ...data, author: user });
+    await this.ideaRepository.save(idea);
+    return this.toResponseObject(idea);
+  }
+
+  async read(id: string): Promise<IdeaRO> {
+    const idea = await this.ideaRepository.findOne({
+      where: { id },
+      relations: ['author'],
+    });
     if (!idea) {
       throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     }
+    return this.toResponseObject(idea);
+  }
+
+  async update(
+    id: string,
+    userId: string,
+    data: Partial<IdeaDTO>,
+  ): Promise<IdeaRO> {
+    let idea = await this.ideaRepository.findOne({
+      where: { id },
+      relations: ['author'],
+    });
+    if (!idea) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+    this.ensureOwnership(idea, userId);
     await this.ideaRepository.update({ id }, data);
-    return await this.ideaRepository.findOne({ where: { id } });
+    idea = await this.ideaRepository.findOne({
+      where: { id },
+      relations: ['author'],
+    });
+    return this.toResponseObject(idea);
   }
 
-  async destroy(id: string) {
-    const idea = await this.ideaRepository.findOne({ where: { id } });
+  async destroy(id: string, userId: string) {
+    const idea = await this.ideaRepository.findOne({
+      where: { id },
+      relations: ['author'],
+    });
     if (!idea) {
       throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     }
+    this.ensureOwnership(idea, userId);
     await this.ideaRepository.delete({ id });
-    return { deleted: true };
+    return this.toResponseObject(idea);
   }
 }
